pyproject.toml
[project]
name = "sigpro-backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asgiref==3.8.1",
    "channels>=4.2.2",
    "channels-redis>=4.2.1",
    "dj-database-url==2.3.0",
    "django==5.1.6",
    "django-debug-toolbar>=6.0.0",
    "django-extensions==3.2.3",
    "django-mathfilters>=1.0.0",
    "gevent>=25.9.1",
    "gunicorn>=23.0.0",
    "ipdb>=0.13.13",
    "ipython>=9.6.0",
    "openpyxl>=3.1.5",
    "pandas>=2.3.3",
    "psycopg2-binary>=2.9.11",
    "pusher>=3.3.3",
    "pyjwt>=2.10.1",
    "sqlparse==0.5.3",
    "typing-extensions==4.12.2",
    "tzdata==2025.1",
]
